+++ app/src/main/java/woowacourse/movie/movie/activity/SeatSelectionActivity.kt
@@ -1,7 +1,6 @@
 package woowacourse.movie.movie.activity
 
 import android.app.AlarmManager
-import android.app.PendingIntent
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
@@ -14,8 +13,7 @@ import domain.Seats
 import woowacourse.movie.R
 import woowacourse.movie.databinding.ActivitySeatSelectionBinding
 import woowacourse.movie.movie.AlarmReceiver
-import woowacourse.movie.movie.AlarmReceiver.Companion.ALARM_CODE
-import woowacourse.movie.movie.AlarmReceiver.Companion.REQUEST_CODE
+import woowacourse.movie.movie.PendingIntentBuilder
 import woowacourse.movie.movie.dto.BookingHistoryDto
 import woowacourse.movie.movie.dto.movie.BookingMovieDto
 import woowacourse.movie.movie.dto.movie.MovieDateDto
@@ -161,7 +159,7 @@ class SeatSelectionActivity : AppCompatActivity() {
 
     private fun setDateTime(bookingMovie: BookingMovieDto): LocalDateTime {
         val dateTime = LocalDateTime.of(bookingMovie.date.date, bookingMovie.time.time)
-        return dateTime.minusMinutes(30L)
+        return dateTime.minusMinutes(1L)
     }
 
     private fun putAlarm(bookingMovie: BookingMovieDto) {
@@ -170,21 +168,8 @@ class SeatSelectionActivity : AppCompatActivity() {
         alarmManager.setExact(
             AlarmManager.RTC_WAKEUP,
             setDateTime(bookingMovie).atZone(ZoneId.systemDefault()).toInstant().toEpochMilli(),
-            createReceiverPendingIntent(bookingMovie),
+            PendingIntentBuilder(this).createReceiverPendingIntent(Intent(this, AlarmReceiver::class.java), bookingMovie),
         )
-    }
-
-    private fun createReceiverPendingIntent(bookingMovie: BookingMovieDto): PendingIntent {
-        return Intent(this, AlarmReceiver::class.java).let {
-            it.action = ALARM_CODE
-            it.putExtra(BOOKING_MOVIE_KEY, bookingMovie)
-            PendingIntent.getBroadcast(
-                this,
-                REQUEST_CODE,
-                it,
-                PendingIntent.FLAG_IMMUTABLE,
-            )
-        }
     }
 
     companion object {
