+++ app/src/main/java/woowacourse/movie/movie/dto/seat/SeatDto.kt
@@ -1,8 +1,10 @@
 package woowacourse.movie.movie.dto.seat
 
-import java.io.Serializable
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
 
-data class SeatDto(val row: SeatRowDto, val col: SeatColDto) : Serializable {
+@Parcelize
+data class SeatDto(val row: SeatRowDto, val col: SeatColDto) : Parcelable {
     fun getString(): String {
         return "${row.row}${col.col}"
     }
